create separate components folder
