+++ src/main/java/com/google/graphgeckos/dashboard/components/Builder.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.graphgeckos.dashboard.storage;
+package com.google.graphgeckos.dashboard.components;
 
 import java.util.ArrayList;
 import java.util.List;
