+++ src/main/java/com/google/graphgeckos/dashboard/components/GitHubData.java
@@ -1,4 +1,4 @@
-package com.google.graphgeckos.dashboard.storage;
+package com.google.graphgeckos.dashboard.components;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
